Extract Labels into its own module
